requirements.txt
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.3.2
lxml==3.4.0
pushbullet.py==0.5.0
python-magic==0.4.6
requests==2.4.1
six==1.8.0
websocket-client==0.21.0
wsgiref==0.1.2
